fix dartls inserting carriage return in code snippets for neovim.
neovim buffers must not have carriage returns. Even in windows neovim,
and file format is set to DOS, Neovim handles all the crlf line endings
of a file and presents a buffer only with lf. However the windows dartls
snippets include crlf and Neovim lsp is not used to remove the cr from
the included snippets.